Move Suggestions to org.eblocker.server.http.ssl
