+++ eblocker-icapserver/src/main/java/org/eblocker/server/http/controller/impl/SSLControllerImpl.java
@@ -40,7 +40,6 @@ import org.eblocker.server.common.page.PageContextStore;
 import org.eblocker.server.common.session.Session;
 import org.eblocker.server.common.session.SessionStore;
 import org.eblocker.server.common.squid.SquidWarningService;
-import org.eblocker.server.common.squid.Suggestions;
 import org.eblocker.server.common.ssl.SslCertificateClientInstallationTracker;
 import org.eblocker.server.common.ssl.SslService;
 import org.eblocker.server.http.controller.SSLController;
@@ -52,6 +51,7 @@ import org.eblocker.server.http.service.ParentalControlService;
 import org.eblocker.server.http.service.SSLWhitelistService;
 import org.eblocker.server.http.service.UserAgentService;
 import org.eblocker.server.http.service.UserService;
+import org.eblocker.server.http.ssl.Suggestions;
 import org.restexpress.Request;
 import org.restexpress.Response;
 import org.restexpress.exception.BadRequestException;
