+++ eblocker-icapserver/src/test/java/org/eblocker/server/http/service/FailedConnectionSuggestionServiceTest.java
@@ -4,8 +4,8 @@ import com.google.common.collect.Sets;
 import org.eblocker.server.common.data.Device;
 import org.eblocker.server.common.squid.FailedConnection;
 import org.eblocker.server.common.squid.SquidWarningService;
-import org.eblocker.server.common.squid.Suggestions;
 import org.eblocker.server.http.ssl.AppWhitelistModule;
+import org.eblocker.server.http.ssl.Suggestions;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
